Update SevenSegment Display documentation
Fixed the sample so it compiled and updated it for common cathode. The diagram is wiring the common pins to ground so I think this should be a common cathode seven segment display. I changed the code to reflect that as well as the documentation. It looks like the TEA5767 is a radio so I changed the reference to what I tested this on, a common cathode 5161AS.