Merge cell doc fixes with latest cell docs structure (#712)
This is a PR that would update the cell docs and fix branch's changes with the structure changes from the cell docs split. It's saying the merge won't work automatically, so I might need to investigate it further.

* Update index.md

Added a step to the Visual Studio 2022 instructions that was missing, but was in the VS Code section. It's the step to install the Meadow Templates.

* Change rule field order and sort rules

* Add redirect for old Deploying_Meadow guide

* Rename original redirect file

* Trying non-encoded redirect destination

* Split up cellular docs into sections

* Add cellular sections to TOC

* Slight tweak from review

* Link together the split cellular pages

---------

Co-authored-by: Al Rodriguez <ProgrammerAl@Outlook.com>
Co-authored-by: adunndevster2 <38018663+adunndevster2@users.noreply.github.com>
Co-authored-by: Jorge Ramirez <ramirez.jorgea@outlook.com>